Tool added

+++ src/main/java/com/factory/entities/Tool.java
@@ -7,7 +7,7 @@ public class Tool {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer Id;
+    private Long Id;
 
     private String Spec;
 
@@ -23,11 +23,11 @@ public class Tool {
     public Tool(){
     }
 
-    public Integer getId() {
+    public Long getId() {
         return Id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         Id = id;
     }
